add simpler toggles
